% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0reg.R
\name{l0reg}
\alias{l0reg}
\title{Solve $l_0$-regularized least squres with Single Best Replacement (SBR) algorithm}
\usage{
l0reg(X, y, lambda, control = list())
}
\arguments{
\item{X}{A n by p numeric design matrix}

\item{y}{A n vector of numeric response}

\item{lambda}{A non-negative numeric penalty parameter}
}
\description{
This is the main function for solving $l_0$-regularized least squres for a given
penalty parameter, based on the Single Best Replacement (SBR) algorithm originally proposed by Soussen et al (2011)
}
